refactor: move category type inside corresponding feature folder
