+++ packages/frontend/features/categories/components/categories-table.tsx
@@ -2,7 +2,7 @@
 import { DataTable } from '@/components/table/data-table'
 import { columns } from '@/features/categories/components/colums'
 import useDeleteCategories from '@/features/categories/hooks/use-delete-categories'
-import { Category } from '@/types/types'
+import { Category } from '@/features/categories/types'
 
 export default function CategoriesTable({
 	categories
