+++ packages/frontend/features/categories/types.d.ts
@@ -1,0 +1,6 @@
+export type Category = {
+	id: string
+	plaidId: string | null
+	name: string
+	userId: string
+}
