+++ packages/frontend/types/types.d.ts
@@ -6,13 +6,6 @@ export type User = {
 	password: string
 }
 
-export type Category = {
-	id: string
-	plaidId: string | null
-	name: string
-	userId: string
-}
-
 export type Transaction = {
 	id: string
 	accountId: string
